requirements.txt
torch==1.12.1
torchvision==0.13.1
torchaudio==0.12.1
transformers==4.28.1
habitat-lab==2.0.0
numpy==1.21.5
opencv-python==4.5.5.64
pyyaml==6.0
tensorrt==8.4.1.5
pycuda==2021.1